#283 move: 세부통계 코드 이동
