+++ src/pages/DashboardPage/DashboardStatisticPage/MajorChart.jsx
